Refactor: Create scroll component to seperate items transformation needed by horizontal scroll
